clean and finish
